[Php74] Move ReservedFnFunctionRector from Transform to Php74 namespace, change fn() to f()
