--- rules/Transform/Rector/Function_/ReservedFnFunctionRector.php
@@ -1,123 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Rector\Transform\Rector\Function_;
-
-use PhpParser\Node;
-use PhpParser\Node\Expr\FuncCall;
-use PhpParser\Node\Identifier;
-use PhpParser\Node\Name;
-use PhpParser\Node\Stmt\Function_;
-use Rector\Core\Contract\Rector\ConfigurableRectorInterface;
-use Rector\Core\Rector\AbstractRector;
-use Rector\Core\ValueObject\PhpVersionFeature;
-use Rector\VersionBonding\Contract\MinPhpVersionInterface;
-use Symplify\RuleDocGenerator\ValueObject\CodeSample\ConfiguredCodeSample;
-use Symplify\RuleDocGenerator\ValueObject\RuleDefinition;
-use Webmozart\Assert\Assert;
-
-/**
- * @changelog https://github.com/php/php-src/pull/3941/files#diff-7e3a1a5df28a1cbd8c0fb6db68f243da
- * @see \Rector\Tests\Transform\Rector\Function_\ReservedFnFunctionRector\ReservedFnFunctionRectorTest
- */
-final class ReservedFnFunctionRector extends AbstractRector implements ConfigurableRectorInterface, MinPhpVersionInterface
-{
-    /**
-     * @api
-     * @deprecated
-     * @var string
-     */
-    final public const RESERVED_NAMES_TO_NEW_ONES = '$reservedNamesToNewOnes';
-
-    /**
-     * @var string[]
-     */
-    private array $reservedNamesToNewOnes = [];
-
-    public function provideMinPhpVersion(): int
-    {
-        return PhpVersionFeature::RESERVED_FN_FUNCTION_NAME;
-    }
-
-    public function getRuleDefinition(): RuleDefinition
-    {
-        return new RuleDefinition('Change fn() function name, since it will be reserved keyword', [
-            new ConfiguredCodeSample(
-                <<<'CODE_SAMPLE'
-class SomeClass
-{
-    public function run()
-    {
-        function fn($value)
-        {
-            return $value;
-        }
-
-        fn(5);
-    }
-}
-CODE_SAMPLE
-                ,
-                <<<'CODE_SAMPLE'
-class SomeClass
-{
-    public function run()
-    {
-        function f($value)
-        {
-            return $value;
-        }
-
-        f(5);
-    }
-}
-CODE_SAMPLE
-                ,
-                [
-                    'fn' => 'someFunctionName',
-                ]
-            ),
-        ]);
-    }
-
-    /**
-     * @return array<class-string<Node>>
-     */
-    public function getNodeTypes(): array
-    {
-        return [Function_::class, FuncCall::class];
-    }
-
-    /**
-     * @param Function_|FuncCall $node
-     */
-    public function refactor(Node $node): ?Node
-    {
-        foreach ($this->reservedNamesToNewOnes as $reservedName => $newName) {
-            if (! $this->isName($node->name, $reservedName)) {
-                continue;
-            }
-
-            $node->name = $node instanceof FuncCall ? new Name($newName) : new Identifier($newName);
-
-            return $node;
-        }
-
-        return null;
-    }
-
-    /**
-     * @param mixed[] $configuration
-     */
-    public function configure(array $configuration): void
-    {
-        $reservedNamesToNewOnes = $configuration[self::RESERVED_NAMES_TO_NEW_ONES] ?? $configuration;
-
-        Assert::allString(array_keys($reservedNamesToNewOnes));
-        Assert::allString($reservedNamesToNewOnes);
-
-        /** @var array<string, string> $reservedNamesToNewOnes */
-        $this->reservedNamesToNewOnes = $reservedNamesToNewOnes;
-    }
-}
